refactor page object to appropriate source folder / package
